Enhance uninstall

+++ scripts/docker.sh
@@ -2,16 +2,14 @@
 
 DOCKER_VERSION="17.03.2"
 BIN_DIR=/usr/local/bin
+DOCKER_SERVICE=/etc/systemd/system/docker.service
 
 install_docker() {
     local download_url="https://download.docker.com/linux/static/stable/x86_64/docker-$DOCKER_VERSION-ce.tgz"
 
     wget -qO- $download_url | tar -xvz -C $BIN_DIR --strip-components=1
 
-    local TEMPLATE=/etc/systemd/system/docker.service
-    if [ ! -f $TEMPLATE ]; then
-        echo "TEMPLATE: $TEMPLATE"
-        cat << EOF > $TEMPLATE
+    cat << EOF > ${DOCKER_SERVICE}
 [Unit]
 Description=Docker Application Container Engine
 Documentation=https://docs.docker.com/engine/reference/commandline/dockerd/
@@ -24,7 +22,7 @@ RestartSec=5
 [Install]
 WantedBy=multi-user.target
 EOF
-    fi
+
     # Enable and start the service
     systemctl daemon-reload
     systemctl enable docker
@@ -37,10 +35,12 @@ EOF
 }
 
 uninstall_docker() {
-    systemctl stop docker
-    systemctl daemon-reload
-    rm /etc/systemd/system/docker.service
-    rm $BIN_DIR/docker*
+    if [ -f ${DOCKER_SERVICE} ]; then
+        systemctl stop docker
+        systemctl daemon-reload
+        rm /etc/systemd/system/docker.service
+    fi
+    rm -f $BIN_DIR/docker*
 }
